Move grandpa pallet to `bridge_common_config.rs`
